объединила всю работу с криптосистемой в один файл
