+++ lab_3/lab3/main.py
@@ -1,8 +1,9 @@
 import argparse
 import const
+import os
 import sys
 
-from crypto_manager import CryptoManager
+from hybrid_crypto_system import CryptoManager
 
 
 def setup_arg_parser():
