+++ lab_3/lab3/public_key.txt
@@ -1,9 +1,9 @@
 -----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAwigJX/p9RO88tL95Zlyj
-4avrRG3AbgJJ8svFiH1ljqE4FeqAWTOF1OQztYYGBsNni3e0CYlMic8soWWRLTA1
-alLuJifBZikstL/y64mHnfwl/jroKmBdBcoRPaZHvuincWPlb10ORP0VCovK9i/9
-vy/yDSCdPjjW0tQPzNye7OmcEl97ozOeWm+185w1+0juV3VUmgG0wL5PIXEH0yNc
-yZRQiKO0cLHjVWSlhENBAHyA23VmGlUacuqpIY8c/ztk7asLfzlaOmC0z7OmZnuy
-VJdIysZ65GDaYYGSzbXs7ljjNkH1mKRi+wf7AZv+V/GLV3GaKMU3ikQ63HCuOqqX
-gQIDAQAB
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAxHo5lfbUYiYqPyJLMs/g
+Lf0W857EvpyS+T+aW3tI8I0kJ4J04F7nfdS5ltz5LENziM1Knf3skKE9BL79iVt7
+JunZz2T+HD/9s2i6widivfGYI5gIQj5ql0KO96nhRvdKxZT5airuzf5kN+nOTg5v
+QnUyrFLPxS23xn4cWusZFhHo6rT/P8xJyP1vui97h454MRWyGz/InZhFrhbehYbj
+zU8k9m4YP5zouwaUnuE7ei1bQancGWa7E7z2aPX2F7r9Q04MyTsF3368MMaY5lBO
+mMft5CqVJX9PyYrx7QVYy9twgXAiQDU+IOgBwLcNy/dQ0R5IB6CNtQSgTeD5PbQX
+EwIDAQAB
 -----END PUBLIC KEY-----
